Replace soong config list variable with selects
It would be more work to support setting configurable properties with
soong config list variables, so instead, replace it with selects.

The selects are more verbose than the list variable, but at the same
time, they're replacing this logic in make:
https://cs.android.com/android/platform/superproject/main/+/main:build/make/core/config.mk;l=870;drc=b20aee1b1e015171986afa5c2ddfcd3452d3b2dc

In the future we may want to come up with a proper list comprehension
syntax in soong to make this shorter.

Bug: 347605145
Bug: 342006386
Test: m nothing --no-skip-soong-tests
Change-Id: I0157056457d8f56c3bba85bd89c682d485d39793
Signed-off-by: Dmitrii <bankersenator@gmail.com>